+++ Cuestionario/Cuestionario/Questionnaire.Handlers/Handlers/Interfaces/IUserAnswerHandler.cs
@@ -4,6 +4,6 @@ namespace Questionnaire.Handlers.Handlers.Interfaces
 {
     public interface IUserAnswerHandler : IBaseHandler
     {
-        int SaveUserAnswer(UserAnswerCreationData pUserAnswer, int pAnswerSessionId);
+        long SaveUserAnswer(UserAnswerCreationData pUserAnswer, long pAnswerSessionId);
     }
 }
